Trying to  fix the search agent

+++ src/pacman/game/Game.java
@@ -1294,7 +1294,7 @@ public final class Game
 	 */
 	public int getManhattanDistance(int fromNodeIndex,int toNodeIndex)
 	{
-		return (int)(Math.abs(currentMaze.graph[fromNodeIndex].x-currentMaze.graph[toNodeIndex].x)+Math.abs(currentMaze.graph[fromNodeIndex].y-currentMaze.graph[toNodeIndex].y));
+		return Math.abs(currentMaze.graph[fromNodeIndex].x-currentMaze.graph[toNodeIndex].x)+Math.abs(currentMaze.graph[fromNodeIndex].y-currentMaze.graph[toNodeIndex].y);
 	}
 	
 	/**
